# Perf: optimize deepks functions cal_f_delta and cal_pdm