--- a/source/module_base/intarray.cpp
+++ b/source/module_base/intarray.cpp
@@ -22,7 +22,7 @@ IntArray::IntArray(const int d1,const int d2)
 	bound3 = bound4 = bound5 = bound6 = 0;
 	size = bound1 * bound2;
 	ptr = new int[size];zero_out();
-	assert( ptr != 0);
+	assert( ptr != nullptr);
 	++arrayCount;
 }
 
@@ -36,7 +36,7 @@ IntArray::IntArray(const int d1,const int d2,const int d3)
 	//set_new_handler(IntArrayAlloc);
 	size = bound1 * bound2 * bound3 ;	//* sizeof(float);
 	ptr = new int[size];zero_out();
-	assert(ptr != 0);
+	assert(ptr != nullptr);
 	++arrayCount;
 }
 
@@ -51,7 +51,7 @@ IntArray::IntArray(const int d1,const int d2,const int d3,const int d4)
 	//set_new_handler(IntArrayAlloc);
 	size = bound1 * bound2 * bound3 * bound4 ;	//* sizeof(float);
 	ptr = new int[size];zero_out();
-	assert(ptr != 0);
+	assert(ptr != nullptr);
 	++arrayCount;
 }
 
@@ -67,7 +67,7 @@ IntArray::IntArray(const int d1,const int d2,const int d3,
 	//set_new_handler(IntArrayAlloc);
 	size = bound1 * bound2 * bound3 * bound4 * bound5;
 	ptr = new int[size];zero_out();
-	assert(ptr != 0);
+	assert(ptr != nullptr);
 	++arrayCount;
 }
 
@@ -84,7 +84,7 @@ IntArray::IntArray(const int d1,const int d2,const int d3,
     //set_new_handler(IntArrayAlloc);
     size = bound1 * bound2 * bound3 * bound4 * bound5 * bound6;
 	ptr = new int[size];zero_out();
-	assert(ptr != 0);
+	assert(ptr != nullptr);
 	++arrayCount;
 }
 
@@ -98,10 +98,10 @@ IntArray ::~IntArray()
 
 void IntArray::freemem()
 {
-	if(ptr!=NULL)
+	if(ptr!= nullptr)
 	{
 		delete [] ptr;
-		ptr = NULL;
+		ptr = nullptr;
 	}
 }
 
@@ -111,7 +111,7 @@ void IntArray::create(const int d1,const int d2,const int d3,const int d4,const 
 	dim = 6;
 	bound1 = d1;bound2 = d2;bound3 = d3;bound4 = d4;bound5 = d5;bound6 = d6;
 	delete[] ptr; ptr = new int[size];
-	assert(ptr != 0);zero_out();
+	assert(ptr != nullptr);zero_out();
 }
 
 void IntArray::create(const int d1,const int d2,const int d3,const int d4,const int d5)
@@ -120,7 +120,7 @@ void IntArray::create(const int d1,const int d2,const int d3,const int d4,const 
 	dim = 5;
 	bound1 = d1;bound2 = d2;bound3 = d3;bound4 = d4;bound5 = d5;
 	delete[] ptr; ptr = new int[size];
-	assert(ptr != 0);zero_out();
+	assert(ptr != nullptr);zero_out();
 }
 
 void IntArray::create(const int d1,const int d2,const int d3,const int d4)
@@ -129,7 +129,7 @@ void IntArray::create(const int d1,const int d2,const int d3,const int d4)
 	dim = 4;
 	bound1 = d1;bound2 = d2;bound3 = d3;bound4 = d4;
 	delete[] ptr; ptr = new int[size];
-	assert(ptr != 0);zero_out();
+	assert(ptr != nullptr);zero_out();
 }
 
 void IntArray::create(const int d1,const int d2,const int d3)
@@ -138,7 +138,7 @@ void IntArray::create(const int d1,const int d2,const int d3)
 	dim = 3;
 	bound1 = d1;bound2 = d2;bound3 = d3;bound4 = 1;
 	delete [] ptr;ptr = new int[size];
-	assert(ptr != 0);zero_out();
+	assert(ptr != nullptr);zero_out();
 }
 
 void IntArray::create(const int d1, const int d2)
@@ -147,7 +147,7 @@ void IntArray::create(const int d1, const int d2)
 	dim = 2;
 	bound1 = d1;bound2 = d2;bound3 = bound4 = 1;
 	delete[] ptr;ptr = new int[size];
-	assert(ptr !=0 );zero_out();
+	assert(ptr != nullptr );zero_out();
 }
 
 //****************************
@@ -155,8 +155,14 @@ void IntArray::create(const int d1, const int d2)
 //****************************
 void IntArray::zero_out(void)
 {
-	if (size <= 0) return;
-	for (int i = 0;i < size; i++) ptr[i] = 0;
+	if (size <= 0) 
+	{
+		return;
+	}
+	for (int i = 0;i < size; i++)
+	{
+		ptr[i] = 0;
+	}
 	return;
 }