--- a/source/module_hamilt_lcao/module_hcontainer/base_matrix.cpp
+++ b/source/module_hamilt_lcao/module_hcontainer/base_matrix.cpp
@@ -131,16 +131,6 @@ assert(this->value_begin != nullptr);
 }
 
 template <typename T>
-T& BaseMatrix<T>::get_value(const size_t& i_row, const size_t& j_col) const
-{
-#ifdef __DEBUG
-assert(this->value_begin != nullptr);
-#endif
-    int index = i_row * this->ncol_local + j_col;
-    return value_begin[index];
-}
-
-template <typename T>
 T* BaseMatrix<T>::get_pointer() const
 {
     return value_begin;