### Linked Issue
Fix #4247

### What's changed?
- add openmp support to these two functions
- inline some hotspot functions in `cal_f_delta` and `cal_pdm`

### Perf Improvement
test case: https://github.com/deepmodeling/abacus-develop/files/15456955/cusolver_mpi_openmp.zip
test command: OMP_NUM_THREADS=6 mpirun -n 2 abacus
||cal_f_delta|cal_pdm|total|
|---|---|---|---|
|before|658.98s|50.09s|815.75s|
|after inline functions|205.87s|29.56s|343.35s|
|after inline and add openmp|28.07s|4.97s|140.46s|

One lesson learned from issue #4247 is that some simple functions are best placed in header files to facilitate compiler inlining, which can have a significant impact on performance.